// workshop/tools/ui/cf-app/app/src/api2/api.ts
export class DeviceDto extends EventEmitter {
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    id?: number;
    /**
     *
     * @type {boolean}
     * @memberof DeviceDto
     */
    isActive?: boolean;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    fields?: string;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    user?: number;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    project?: number;
    /**
     *
     * @type {number}
     * @memberof DeviceDto
     */
    flags?: number;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    state?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    host?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    port?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    protocol?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    logging?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    retry?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    blocks?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    commandSettings?: string;
    /**
     *
     * @type {string}
     * @memberof DeviceDto
     */
    settings?: string;
}
// ...
export interface InDeviceDto {
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    id?: number;
    /**
     *
     * @type {boolean}
     * @memberof InDeviceDto
     */
    isActive?: boolean;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    fields?: string;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    user?: number;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    project?: number;
    /**
     *
     * @type {number}
     * @memberof InDeviceDto
     */
    flags?: number;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    state?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    host?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    port?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    protocol?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    logging?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    retry?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    blocks?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    commandSettings?: string;
    /**
     *
     * @type {string}
     * @memberof InDeviceDto
     */
    settings?: string;
}